changed return type

+++ CodeWars/8KYU/AllStarCodeChallenge18/Java/sol3.java
@@ -1,7 +1,7 @@
 import java.util.stream.*;
 
 public class CodeWars {
-  public static long strCount(String str, char letter) {
+  public static int strCount(String str, char letter) {
     return str.chars().filter(c -> c == letter).count();
   }
 }
